11/6 Meeting
We finally understand whats going on!
Just need to take value out of a swerve module state (they are marked as public). Each swerve module state is stored in an array which I have no renamed "swerveModuleStatesArray" and that's what the kinematics class outputs. We take these values (which are exactly what we need) and put them into the PID (replacing that one placeholder). Then  the only thing left will be the controls (which should output as speeds and replace the tempXTarget, tempYTarget, tempRotationTarget, tempRotation2d which are in the creation of the swerve module state array and need to be replaced with the target speeds we get from the controller. Then after some bits and bobs we should be ready.

Co-authored-by: Bry <BCODERW@users.noreply.github.com>